03_notes_service/requirements.txt
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
